+++ example-react-app/src/app/associationM2M-management/AssociationM2MBrowserTable.tsx
@@ -66,6 +66,7 @@ const AssociationM2MBrowserTable = observer(
       handlePaginationChange,
       handleDeleteBtnClick,
       handleCreateBtnClick,
+      handleCloneBtnClick,
       handleEditBtnClick,
       goToParentScreen,
       entityListState
@@ -77,7 +78,8 @@ const AssociationM2MBrowserTable = observer(
       onEntityListChange,
       onPagination: saveHistory,
       onEntityDelete,
-      onOpenScreenError
+      onOpenScreenError,
+      lazyLoading: true
     });
 
     const selectEntityHandler = useCallback(() => {
@@ -153,6 +155,21 @@ const AssociationM2MBrowserTable = observer(
         </EntityPermAccessControl>,
         <EntityPermAccessControl
           entityName={ENTITY_NAME}
+          operation="create"
+          key="create"
+        >
+          <Button
+            htmlType="button"
+            style={{ margin: "0 12px 12px 0" }}
+            disabled={entityListState.selectedEntityId == null}
+            type="default"
+            onClick={handleCloneBtnClick}
+          >
+            <FormattedMessage id="common.clone" />
+          </Button>
+        </EntityPermAccessControl>,
+        <EntityPermAccessControl
+          entityName={ENTITY_NAME}
           operation="update"
           key="update"
         >
@@ -203,6 +220,7 @@ const AssociationM2MBrowserTable = observer(
 
     return (
       <DataTable
+        tableId={ROUTING_PATH + ENTITY_NAME}
         items={items}
         count={count}
         relationOptions={relationOptions}
@@ -220,6 +238,7 @@ const AssociationM2MBrowserTable = observer(
         onPaginationChange={handlePaginationChange}
         hideSelectionColumn={true}
         buttons={buttons}
+        executeListQuery={entityList == null ? executeListQuery : undefined}
       />
     );
   }