+++ example-react-app/src/app/car-master-detail/CarMasterDetailBrowser.tsx
@@ -100,7 +100,8 @@ const CarMasterDetailBrowser = observer((props: EntityListProps<Car>) => {
     onEntityListChange,
     onPagination: saveHistory,
     onEntityDelete,
-    onOpenScreenError
+    onOpenScreenError,
+    lazyLoading: true
   });
 
   if (error != null) {
@@ -162,6 +163,7 @@ const CarMasterDetailBrowser = observer((props: EntityListProps<Car>) => {
 
   return (
     <DataTable
+      tableId={ROUTING_PATH + ENTITY_NAME}
       items={items}
       count={count}
       relationOptions={relationOptions}
@@ -193,6 +195,7 @@ const CarMasterDetailBrowser = observer((props: EntityListProps<Car>) => {
       onPaginationChange={handlePaginationChange}
       hideSelectionColumn={true}
       buttons={buttons}
+      executeListQuery={entityList == null ? executeListQuery : undefined}
     />
   );
 });