+++ example-react-app/src/app/car-multi-selection-table/CarMultiSelectionTable.tsx
@@ -100,7 +100,8 @@ const CarMultiSelectionTable = observer((props: EntityListProps<Car>) => {
     onEntityListChange,
     onPagination: saveHistory,
     onEntityDelete,
-    onOpenScreenError
+    onOpenScreenError,
+    lazyLoading: true
   });
 
   if (error != null) {
@@ -148,6 +149,7 @@ const CarMultiSelectionTable = observer((props: EntityListProps<Car>) => {
 
   return (
     <DataTable
+      tableId={ROUTING_PATH + ENTITY_NAME}
       items={items}
       count={count}
       relationOptions={relationOptions}
@@ -179,6 +181,7 @@ const CarMultiSelectionTable = observer((props: EntityListProps<Car>) => {
       onSortOrderChange={handleSortOrderChange}
       onPaginationChange={handlePaginationChange}
       buttons={buttons}
+      executeListQuery={entityList == null ? executeListQuery : undefined}
     />
   );
 });