+++ example-react-app/src/app/weird-string-id-browser-table/WeirdStringIdBrowserTable.tsx
@@ -67,6 +67,7 @@ const WeirdStringIdBrowserTable = observer(
       handlePaginationChange,
       handleDeleteBtnClick,
       handleCreateBtnClick,
+      handleCloneBtnClick,
       handleEditBtnClick,
       goToParentScreen,
       entityListState
@@ -78,7 +79,8 @@ const WeirdStringIdBrowserTable = observer(
       onEntityListChange,
       onPagination: saveHistory,
       onEntityDelete,
-      onOpenScreenError
+      onOpenScreenError,
+      lazyLoading: true
     });
 
     const selectEntityHandler = useCallback(() => {
@@ -154,6 +156,21 @@ const WeirdStringIdBrowserTable = observer(
         </EntityPermAccessControl>,
         <EntityPermAccessControl
           entityName={ENTITY_NAME}
+          operation="create"
+          key="create"
+        >
+          <Button
+            htmlType="button"
+            style={{ margin: "0 12px 12px 0" }}
+            disabled={entityListState.selectedEntityId == null}
+            type="default"
+            onClick={handleCloneBtnClick}
+          >
+            <FormattedMessage id="common.clone" />
+          </Button>
+        </EntityPermAccessControl>,
+        <EntityPermAccessControl
+          entityName={ENTITY_NAME}
           operation="update"
           key="update"
         >
@@ -204,6 +221,7 @@ const WeirdStringIdBrowserTable = observer(
 
     return (
       <DataTable
+        tableId={ROUTING_PATH + ENTITY_NAME}
         items={items}
         count={count}
         relationOptions={relationOptions}
@@ -221,6 +239,7 @@ const WeirdStringIdBrowserTable = observer(
         onPaginationChange={handlePaginationChange}
         hideSelectionColumn={true}
         buttons={buttons}
+        executeListQuery={entityList == null ? executeListQuery : undefined}
       />
     );
   }