+++ packages/jmix-react-web/src/crud/editor/useEntityEditor.ts
@@ -215,7 +215,8 @@ export function useEntityEditor<
     loadQueryOptions,
     entityInstance,
     entityId,
-    entityName
+    entityName,
+    cloneEntity: multiScreen?.params?.cloneEntity
   });
 
   useEntityEditorForm(item, entityName);
@@ -236,7 +237,9 @@ export function useEntityEditor<
     listQueryName,
     entityName,
     goToParentScreen,
-    entityId,
+    entityId: multiScreen?.params?.cloneEntity
+      ? undefined
+      : entityId,
     entityInstance,
     onCommit,
     uiKit_to_jmixFront,