+++ packages/jmix-addon-charts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/Haulmont/jmix-frontend/compare/@haulmont/jmix-addon-charts@0.1.0-next.4...@haulmont/jmix-addon-charts@3.0.0) (2022-01-11)
+
+**Note:** Version bump only for package @haulmont/jmix-addon-charts
+
+
+
+
+
 # [0.1.0-next.5](https://github.com/Haulmont/jmix-frontend/compare/@haulmont/jmix-addon-charts@0.1.0-next.4...@haulmont/jmix-addon-charts@0.1.0-next.5) (2021-12-22)